% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shijuan_fenxi.R
\name{shijuan_fenxi}
\alias{shijuan_fenxi}
\title{试卷分析}
\usage{
shijuan_fenxi(file, ques_cols = 2:5, fenZhi, sheetname = 1)
}
\arguments{
\item{file}{xlsx格式的成绩表，第一列为学号或姓名，中间列为个小题得分，最后一列为总分；每行为一个学生分数；文件第一行为各列名字（取什么名字无影响）；}

\item{ques_cols}{小题放置的列数，如2:5，表示从第2列到5列；}

\item{fenZhi}{小题按顺序的总分放在 c()中，如,c(10,20,30,40)}

\item{sheetname}{工作表序号，默认为第1个。如果数字房子第一工作表中，不要设置。}
}
\value{
results 各种结果和图表
}
\description{
试卷分析
}
\examples{
data = file.path(system.file(package = "glantools"),"test.xlsx")
shijuan_fenxi(data,ques_cols=2:5,fenZhi=c(10,20,30,40))
data2 = file.path(system.file(package = "glantools"),"test2.xlsx")
shijuan_fenxi(data2,ques_cols=2:5,fenZhi=c(30,15,30,25))

}
\author{
lgm
}
